Move SortRenderer into context.js
